rename routePrefix to prefix, make sure prefix is used in the CLI output
