+++ skills/counterfact-repl/examples.md
@@ -18,7 +18,7 @@ GET http://localhost:3100/_counterfact/api/health
   "status": "ok",
   "port": 3100,
   "basePath": "/",
-  "routePrefix": "/_counterfact/api",
+  "prefix": "/_counterfact/api",
   "uptime": 456.78
 }
 ```
