+++ test/app.test.ts
@@ -17,7 +17,7 @@ const mockConfig = {
   startRepl: false,
   startServer: false,
   watch: { routes: false, types: false },
-  routePrefix: "",
+  prefix: "",
 };
 
 describe("counterfact", () => {
