+++ test/server/dispatcher.test.ts
@@ -1242,7 +1242,7 @@ describe("given a request that contains the differently cased path", () => {
         port: 3100,
         proxyPaths: new Map(),
         proxyUrl: "",
-        routePrefix: "",
+        prefix: "",
         startAdminApi: false,
         startRepl: false,
         startServer: true,
@@ -1401,7 +1401,7 @@ describe("given a request that contains the differently cased path", () => {
         port: 3100,
         proxyPaths: new Map(),
         proxyUrl: "",
-        routePrefix: "",
+        prefix: "",
         startAdminApi: false,
         startRepl: false,
         startServer: true,
